+++ projects/ui-framework/src/lib/lists/tree-list/tree-list-IO.abstract.ts
@@ -8,7 +8,7 @@ import {
   TreeListValue,
   TreeListItemMap,
 } from './tree-list.interface';
-import { BTL_KEYMAP_DEF } from './tree-list.const';
+import { BTL_KEYMAP_DEF, BTL_VALUE_SEPARATOR_DEF } from './tree-list.const';
 import { SelectType } from '../list.enum';
 import { ListFooterActions } from '../list.interface';
 
@@ -23,7 +23,7 @@ export abstract class TreeListInputOutput implements TreeListComponentIO {
   @Input() itemsMap: TreeListItemMap = new Map();
 
   @Input() type: SelectType = SelectType.multi;
-  @Input() valueSeparatorChar = ' / ';
+  @Input() valueSeparatorChar = BTL_VALUE_SEPARATOR_DEF;
   @Input() maxHeightItems = 8;
   @Input() startCollapsed = true;
   @Input() focusOnInit = false;
