+++ projects/ui-framework/src/lib/lists/tree-list/tree-list.const.ts
@@ -22,3 +22,5 @@ export const BTL_KEYMAP_SERVER: TreeListKeyMap = {
   name: 'value',
   children: 'children',
 };
+
+export const BTL_VALUE_SEPARATOR_DEF = ' / ';
