+++ projects/ui-framework/src/lib/lists/tree-list/tree-list.stories.common.ts
@@ -1,6 +1,6 @@
 import { select, number, text, boolean, object } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
-import { BTL_KEYMAP_SERVER } from './tree-list.const';
+import { BTL_KEYMAP_SERVER, BTL_VALUE_SEPARATOR_DEF } from './tree-list.const';
 import { SelectType } from '../list.enum';
 import {
   HListMockSimple,
@@ -20,7 +20,11 @@ export const TreeListStoriesCommonProps = (
   serverKeyMap: BTL_KEYMAP_SERVER,
   type: select('type', Object.values(SelectType), SelectType.single, 'Props'),
   maxHeightItems: number('maxHeightItems', 8, {}, 'Props'),
-  valueSeparatorChar: text('valueSeparatorChar', ' / ', 'Props'),
+  valueSeparatorChar: text(
+    'valueSeparatorChar',
+    BTL_VALUE_SEPARATOR_DEF,
+    'Props'
+  ),
   startCollapsed: boolean('startCollapsed', true, 'Props'),
   readonly: boolean('readonly', false, 'Props'),
   disabled: boolean('disabled', false, 'Props'),
