+++ projects/ui-framework/src/lib/lists/tree-list/tree-list/tree-list.module.ts
@@ -7,6 +7,8 @@ import { SearchModule } from '../../../search/search/search.module';
 import { ListFooterModule } from '../../list-footer/list-footer.module';
 import { TreeListControlsService } from '../services/tree-list-controls.service';
 import { TreeListViewService } from '../services/tree-list-view.service';
+import { TreeListValueService } from '../services/tree-list-value.service';
+import { TreeListSearchService } from '../services/tree-list-search.service';
 
 @NgModule({
   declarations: [TreeListComponent],
@@ -16,6 +18,8 @@ import { TreeListViewService } from '../services/tree-list-view.service';
     TreeListModelService,
     TreeListViewService,
     TreeListControlsService,
+    TreeListValueService,
+    TreeListSearchService,
   ],
 })
 export class TreeListModule {}
